Move new stac api into a library, restore old api
The old dc_to_stac function is used externally so we need to keep
a shim of it for backwards compatibilty.
